+++ src/spaczz/process.py
@@ -2,13 +2,13 @@
 from __future__ import annotations
 
 from itertools import tee
-from typing import Any, Callable, Dict, Iterable
+from typing import Any, Callable, Iterable
 
 from rapidfuzz import fuzz
 from spacy.tokens import Doc
 
 
-def map_chars_to_tokens(doc: Doc) -> Dict[int, int]:
+def map_chars_to_tokens(doc: Doc) -> dict[int, int]:
     """Maps characters in a `Doc` object to tokens."""
     chars_to_tokens = {}
     for token in doc:
@@ -37,7 +37,7 @@ class FuzzyFuncs:
         match_type (str): Whether the fuzzy matching functions
             should support multi-token strings ("phrase") or
             only single-token strings ("token").
-        _fuzzy_funcs (Dict[str, Callable[[str, str], int]]):
+        _fuzzy_funcs (dict[str, Callable[[str, str], int]]):
             The available fuzzy matching functions:
             "simple" = `ratio`
             "partial" = `partial_ratio`
@@ -65,7 +65,7 @@ class FuzzyFuncs:
         """
         self.match_type = match_type
         if match_type == "phrase":
-            self._fuzzy_funcs: Dict[str, Callable[[str, str], int]] = {
+            self._fuzzy_funcs: dict[str, Callable[[str, str], int]] = {
                 "simple": fuzz.ratio,
                 "partial": fuzz.partial_ratio,
                 "token_set": fuzz.token_set_ratio,