+++ src/spaczz/search/similaritysearcher.py
@@ -48,13 +48,13 @@ class SimilaritySearcher(_PhraseSearcher):
                 with spaCy pipelines.
 
         Warnings:
-            UserWarning:
+            MissingVectorsWarning:
                 If vocab does not contain any word vectors.
         """
         super().__init__(vocab=vocab)
         if vocab.vectors.n_keys == 0:
             warnings.warn(
-                """The spaCy Vocab object has no word vectors.\n
+                """The spaCy Vocab object has no word vectors.
                 Similarity results may not be useful.""",
                 MissingVectorsWarning,
             )